requirements.txt
InquirerPy
requests
python_dotenv
